[AARCH64] Reformat inline-asm in elf_machine_load_address.
This patch reformats the inline-asm in elf_machine_load_address so it is
easier to change only part of the inline-asm.  That is using string
concating instead of string continueing.

Also document on why this inline-asm works, it depends on the 32bit
relocation being resolved at link time.

* sysdeps/aarch64/dl-machine.h (elf_machine_load_address):
Refactor inline-asm.  Also add comment.